langgraph_demo/pyproject.toml
[project]
name = "langgraph-dynamic-demo"
version = "1.0.0"
description = "LangGraph 动态图构建演示项目"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
